requirements.txt
faster-whisper>=1.0.0
webrtcvad-wheels>=2.0.10.post2
sounddevice>=0.4.6
pydub>=0.25.1
numpy>=1.24
edge-tts>=6.1.10
piper-tts>=1.2.0
tomli>=2.0.1
tomli-w>=1.0.0
transformers>=4.46.0
sentencepiece>=0.2.0
sacremoses>=0.1.1
requests>=2.31.0
